Let me tell you the story of a hack night. Simpleweb had a hack night in 17th February 2015 focused around pusher. Now, pusher is really awesome but it doesn't quite live up to its name in that it doesn't support push notifications. Jamie and I set out to change this and pushitrealgood was born.
We hacked together this RoR application, an iPhone app and a Ruby script to build an 'in the middle' service that takes your pusher account details and subscribes to an APN channel listening for message events.

Devices register against an account and push messages can be sent to all (or at some point a specific) device. Its all a bit hacked together at some point, and unsurprisingly I didn't find any time commit which is why I am writing this absurdly long commit message right here. I'll clean it up and release it as a service within the next couple of weeks but for now you may peruse through is source code. Ohh and in case you are wondering yes I have revoked those tokens, maybe when I find time I'll move them into the environment but why should that stop me from sharing this code with you now.

Right I better be off. And @pusher thanks for the Drone I <3 it greatly :)

◯  ◯
  ╳
◯  ◯